move ValidationUtil around and keep default-markin in one place
