+++ src/components/moneyflow/EditMoneyflowModal.vue
@@ -103,9 +103,9 @@ import type { Moneyflow } from "@/model/moneyflow/Moneyflow";
 import type { MoneyflowReceipt } from "@/model/moneyflow/MoneyflowReceipt";
 import { mapImportedMoneyflowReceiptToMoneyflowReceipt } from "@/service/mapper/ImportedToMoneyflowReceiptMapper";
 import MoneyflowReceiptService from "@/service/MoneyflowReceiptService";
-import { createFormContext } from "@/service/util/ValidationUtil.ts";
 import { handleBackendError } from "@/tools/views/HandleBackendError";
 import { isDesktop } from "@/tools/views/IsDesktop";
+import { createFormContext } from "@/tools/views/ValidationUtil.ts";
 import { Eye, ReceiptText, Save, Trash2, Undo2 } from "@lucide/vue";
 import { storeToRefs } from "pinia";
 import { computed, ref, toRaw, useTemplateRef, watch } from "vue";
