+++ src/tests/components/moneyflow/EditMoneyflowBase.test.ts
@@ -18,7 +18,6 @@ import ImportedMoneyflowServiceMocker from "@/service/mocker/ImportedMoneyflowSe
 import MoneyflowServiceMocker from "@/service/mocker/MoneyflowServiceMocker";
 import PostingAccountServiceMocker from "@/service/mocker/PostingAccountServiceMocker";
 import PreDefMoneyflowServiceMocker from "@/service/mocker/PreDefMoneyflowServiceMocker";
-import { createFormContext } from "@/service/util/ValidationUtil";
 import { StoreService } from "@/stores/StoreService";
 import {
   assertHaveBeenCalledWith,
@@ -35,6 +34,7 @@ import {
   ToastView,
   ToggleView,
 } from "@/tests/TestViews";
+import { createFormContext } from "@/tools/views/ValidationUtil";
 import "@testing-library/jest-dom/vitest";
 import { render } from "@testing-library/vue";
 import { createPinia, setActivePinia } from "pinia";
