+++ src/views/comparedata/CompareData.vue
@@ -137,9 +137,9 @@ import type { Moneyflow } from "@/model/moneyflow/Moneyflow";
 import type { SelectBoxValue } from "@/model/SelectBoxValue";
 import CompareDataService from "@/service/CompareDataService";
 import MoneyflowService from "@/service/MoneyflowService";
-import { createFormContext } from "@/service/util/ValidationUtil.ts";
 import { useCapitalsourceStore } from "@/stores/CapitalsourceStore";
 import { handleBackendError } from "@/tools/views/HandleBackendError";
+import { createFormContext } from "@/tools/views/ValidationUtil.ts";
 import { globErr } from "@/tools/views/ZodUtil";
 import { Eye } from "@lucide/vue";
 import { computed, onMounted, provide, ref } from "vue";
