+++ src/views/etf/elements/CalcEtfSaleForm.vue
@@ -121,8 +121,8 @@ import type { Etf } from "@/model/etf/Etf";
 import type { EtfSalesCalculation } from "@/model/etf/EtfSalesCalculation";
 import type { EtfSummary } from "@/model/etf/EtfSummary";
 import EtfService from "@/service/EtfService";
-import { createFormContext } from "@/service/util/ValidationUtil.ts";
 import { handleBackendError } from "@/tools/views/HandleBackendError";
+import { createFormContext } from "@/tools/views/ValidationUtil.ts";
 import { amountSchema } from "@/tools/views/ZodUtil";
 import { Euro, Percent } from "@lucide/vue";
 import { ref, watch } from "vue";
