+++ src/views/etf/elements/CreateEtfFlowModal.vue
@@ -104,10 +104,10 @@ import { Button } from "@/components/ui/button";
 import type { EtfFlow } from "@/model/etf/EtfFlow";
 import type { SelectBoxValue } from "@/model/SelectBoxValue";
 import CrudEtfFlowService from "@/service/CrudEtfFlowService";
-import { createFormContext } from "@/service/util/ValidationUtil.ts";
 import { useEtfStore } from "@/stores/EtfStore";
 import { formatTime } from "@/tools/views/FormatDate";
 import { handleBackendError } from "@/tools/views/HandleBackendError";
+import { createFormContext } from "@/tools/views/ValidationUtil";
 import { amountSchema, globErr } from "@/tools/views/ZodUtil";
 import { Euro, Save, Undo2 } from "@lucide/vue";
 import { storeToRefs } from "pinia";
