+++ src/views/etf/elements/CreateEtfModal.vue
@@ -175,7 +175,7 @@ import InputStandard from "@/components/common/InputStandard.vue";
 import ModalVue from "@/components/common/Modal.vue";
 import { Button } from "@/components/ui/button";
 import type { Etf } from "@/model/etf/Etf";
-import { createFormContext } from "@/service/util/ValidationUtil.ts";
+import { createFormContext } from "@/tools/views/ValidationUtil";
 import { amountSchema, globErr } from "@/tools/views/ZodUtil";
 import { Euro, Percent, Save, Star, Undo2 } from "@lucide/vue";
 import { storeToRefs } from "pinia";
