+++ src/views/etf/elements/CreateEtfPreliminaryLumpSumModalPiece.vue
@@ -85,9 +85,9 @@ import type { EtfPreliminaryLumpSum } from "@/model/etf/EtfPreliminaryLumpSum";
 import { EtfPreliminaryLumpSumType } from "@/model/etf/EtfPreliminaryLumpSumType";
 import type { SelectBoxValue } from "@/model/SelectBoxValue";
 import CrudEtfPreliminaryLumpSumService from "@/service/CrudEtfPreliminaryLumpSumService";
-import { createFormContext } from "@/service/util/ValidationUtil.ts";
 import { useEtfStore } from "@/stores/EtfStore";
 import { handleBackendError } from "@/tools/views/HandleBackendError";
+import { createFormContext } from "@/tools/views/ValidationUtil";
 import { amountSchema, globErr } from "@/tools/views/ZodUtil";
 import { Euro, Save, Undo2 } from "@lucide/vue";
 import { storeToRefs } from "pinia";
