+++ src/views/receipt/elements/ImportReceiptsRow.vue
@@ -204,9 +204,9 @@ import { MoneyflowReceiptType } from "@/model/moneyflow/MoneyflowReceiptType";
 import ImportedMoneyflowReceiptService from "@/service/ImportedMoneyflowReceiptService";
 import { mapImportedMoneyflowReceiptToMoneyflowReceipt } from "@/service/mapper/ImportedToMoneyflowReceiptMapper";
 import MoneyflowService from "@/service/MoneyflowService";
-import { createFormContext } from "@/service/util/ValidationUtil.ts";
 import { toFixed } from "@/tools/math";
 import { handleBackendError } from "@/tools/views/HandleBackendError";
+import { createFormContext } from "@/tools/views/ValidationUtil.ts";
 import { amountSchema, globErr } from "@/tools/views/ZodUtil";
 import { Euro, Save, Search } from "@lucide/vue";
 import { computed, inject, nextTick, onMounted, ref } from "vue";
