+++ src/views/user/ChangePassword.vue
@@ -53,7 +53,7 @@
 </template>
 
 <script lang="ts" setup>
-import { createFormContext } from "@/service/util/ValidationUtil.ts";
+import { createFormContext } from "@/tools/views/ValidationUtil";
 import { onMounted, ref, watch } from "vue";
 import { useI18n } from "vue-i18n";
 import { string, ZodType } from "zod";
