+++ src/views/user/elements/CreateUserModal.vue
@@ -196,8 +196,8 @@ import type { User } from "@/model/user/User";
 import { UserRole, userRoleValues } from "@/model/user/UserRole";
 import GroupService from "@/service/GroupService";
 import UserService from "@/service/UserService";
-import { createFormContext } from "@/service/util/ValidationUtil.ts";
 import { handleBackendError } from "@/tools/views/HandleBackendError";
+import { createFormContext } from "@/tools/views/ValidationUtil";
 import { globErr } from "@/tools/views/ZodUtil";
 import { Save, Undo2 } from "@lucide/vue";
 import { storeToRefs } from "pinia";
